TS-1084: Add format string checking
Add format string checking to internal and external APIs that take
a printf(3) format string. No functional changes.

Fix all the resulting warnings
    - we needed a lot of %PRId64 tokens
    - time_t is formatted as long long for portability
    - size_t became %zu
    - pointers all became %p

+++ mgmt/Main.cc
@@ -754,7 +754,7 @@ main(int argc, char **argv)
   //   UI record for this machine
   overviewGenerator->addSelfRecord();
   webThrId = ink_thread_create(webIntr_main, NULL);     /* Spin web agent thread */
-  Debug("lm", "Created Web Agent thread (%d)", webThrId);
+  Debug("lm", "Created Web Agent thread (%lld)", (long long)webThrId);
   lmgmt->listenForProxy();
 
   /* Check the permissions on vip_config */
